group creation alter

+++ src/main/java/com/axioma/axiomatrainee/model/dto/DoneExerciseDto.java
@@ -15,7 +15,7 @@ public class DoneExerciseDto {
 
     DoneExerciseId doneExerciseId;
 
-    Integer score;
+    Double score;
 
     ExerciseType exerciseType;
